requirements.txt
openai
langgraph
langchain
langchain-openai
langchain-community
tavily-python
fastapi
uvicorn
streamlit
requests
python-dotenv
pydantic
reportlab
